requirements.txt
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
future==0.17.1
gym==0.12.0
idna==2.8
kiwisolver==1.0.1
matplotlib==3.0.3
numpy==1.16.2
Pillow==5.4.1
pyglet==1.3.2
pyparsing==2.3.1
python-dateutil==2.8.0
requests==2.21.0
scipy==1.2.1
six==1.12.0
torch==1.0.1.post2
torchvision==0.2.2.post3
urllib3==1.24.1
